src/services/graph_service.rs
pub struct GraphService;

impl GraphService {
    pub fn get_graph_data(state: &AppState) -> Result<GraphData, std::io::Error> {
        // Logic to get graph data
        Ok(GraphData::default())
    }

    pub fn refresh_graph_data(state: &AppState) -> Result<GraphData, std::io::Error> {
        // Logic to refresh graph data
        Ok(GraphData::default())
    }

    pub fn build_edges(state: &AppState) -> Result<Vec<Edge>, std::io::Error> {
        // Logic to build edges
        Ok(Vec::new())
    }
}
